virt-test: Use trace_printk and get rid of extranous printk
It is better to use trace_printk than printk, because printk goes to
host userspace for printing stuff and may skew results in terms of
cache utilization and such.

It is required to use enable CONFIG_FUNCTION_TRACE in the kernel for
seeing this data through /sys/kernel/debugfs/trace.