arm64: KVM: Add exit analysis trap statistics
WARNING: This is a major hack, distustingly tying the KVM code to the
gic code - not for human consumption!

This commit adds a debugfs file under kvm which a user can read from to
see a breakdown of time spent outside a VM and of how much time is spent
in each portion of the host code.

Users can also write 'reset' to the file to reset the stats.  This is
convenient for measuring the behavior under a certain workload.

Whenever an IRQ hits, call the KVM notifier, which calls into the
arch-specific code, which checks if the IRQ hit while we're running a
VCPU, and if we are, then account for the time spent in the interrupt
handler.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>

Conflicts:
	arch/arm/kvm/arm.c
	arch/arm/kvm/mmio.c
	drivers/irqchip/irq-gic.c