Revert "perf callchain: Move cpumode resolve code to add_callchain_ip"
This reverts commit 2e77784bb7d882647c33d8e75a650625e6df0f8b.

The commit states that there's "No change in behavior", which is
unfortunately not correct, because before the commit the code would
sample the initial callchain entry to determine the context and set
cpumode accordingly, and use that cpumode for the remaining callchain
entries.  After the commit, however, each IP in the callchain is sampled
and if the IP is not above the PERF_CONTEXT_MAX magic, then it just
falls back to the default of PERF_RECORD_MISC_USER, which I think only
works for kernel callchains because of the hack for resolving the vdso
symbols.

Cc: Kan Liang <kan.liang@intel.com>
Cc: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>