arm/arm64: KVM: Tweak perf kvm_is_in_guest
Currently we test if we have a running_vcpu set in the kvm_is_in_guest
perf hook function, but this is slightly misrepresenting things in perf
results, because once we come back from the guest, and are doing some
sort of in-vcpu-thread backend processing (for example through en
ioeventfd notification), we will not be scheduled out or have called
kvm_vcpu_put, so we inaccurately count samples for such time towards the
guest.

Instead, take the same approach as the time accounting in the system
otherwise, and check the PF_VCPU flag which is set and cleared by
kvm_guest_enter/exit, respecitively.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>